v2 finalizada

+++ src/main/java/br/com/fiap/espb/checkpoint/receitas/service/IngredienteService.java
@@ -1,11 +1,14 @@
 package br.com.fiap.espb.checkpoint.receitas.service;
 
 import br.com.fiap.espb.checkpoint.receitas.domainModel.Ingrediente;
+import br.com.fiap.espb.checkpoint.receitas.dto.IngredienteDto;
+import br.com.fiap.espb.checkpoint.receitas.mapper.IngredienteMapper;
 import br.com.fiap.espb.checkpoint.receitas.repository.IngredienteRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class IngredienteService {
@@ -19,22 +22,38 @@ public class IngredienteService {
     //criando CRUD
 
     //listar
-    public List<Ingrediente> listarTodos(){
-        return ingredienteRepo.findAll();
+    public List<IngredienteDto> listarTodos(){
+        return ingredienteRepo.findAll().stream().map(IngredienteMapper::toDto).collect(Collectors.toList());
     }
 
     //Busca por id
-    public Optional<Ingrediente> buscarId(int idIngrediente){
-        return ingredienteRepo.findById(idIngrediente);
+    public IngredienteDto buscarId(int idIngrediente){
+        Ingrediente ingrediente = ingredienteRepo.findById(idIngrediente).orElseThrow(() -> new RuntimeException("Ingrediente não encontrado para o id: " + idIngrediente));
+        return IngredienteMapper.toDto(ingrediente);
     }
 
     //Inserindo
-    public Ingrediente inserir(Ingrediente ingrediente){
-        return ingredienteRepo.save(ingrediente);
+    public IngredienteDto inserir(IngredienteDto ingrediente){
+        Ingrediente ingredienteSalvar = IngredienteMapper.toEntity(ingrediente);
+        Ingrediente salvar = ingredienteRepo.save(ingredienteSalvar);
+        return IngredienteMapper.toDto(salvar);
+    }
+
+    //Atualizar
+    public IngredienteDto atualizar(int idIngrediente, IngredienteDto ingredienteDto){
+        Ingrediente ingrediente = ingredienteRepo.findById(idIngrediente).orElseThrow(() -> new RuntimeException("Ingrediente não encontrado para o id: " + idIngrediente));
+        ingrediente.setNome(ingredienteDto.getNome());
+        ingrediente.setQuantidade(ingredienteDto.getQuantidade());
+        ingrediente.setMedida(ingredienteDto.getMedida());
+        Ingrediente atualizada = ingredienteRepo.save(ingrediente);
+        return IngredienteMapper.toDto(atualizada);
     }
 
     //Deletando
     public void deletar(int idIngrediente){
+        if (!ingredienteRepo.existsById(idIngrediente)){
+            throw new RuntimeException("Ingrediente não encontrado para o id: " + idIngrediente);
+        }
         ingredienteRepo.deleteById(idIngrediente);
     }
 }
