+++ src/main/java/br/com/fiap/espb/checkpoint/receitas/service/ReceitasService.java
@@ -1,12 +1,16 @@
 package br.com.fiap.espb.checkpoint.receitas.service;
 
+import br.com.fiap.espb.checkpoint.receitas.domainModel.Ingrediente;
 import br.com.fiap.espb.checkpoint.receitas.domainModel.Receitas;
+import br.com.fiap.espb.checkpoint.receitas.dto.ReceitasDto;
+import br.com.fiap.espb.checkpoint.receitas.mapper.IngredienteMapper;
+import br.com.fiap.espb.checkpoint.receitas.mapper.ReceitasMapper;
 import br.com.fiap.espb.checkpoint.receitas.repository.ReceitasRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.OptionalInt;
+import java.util.stream.Collectors;
 
 @Service
 public class ReceitasService {
@@ -18,24 +22,51 @@ public class ReceitasService {
     }
 
     //criando CRUD
+    public ReceitasDto inserirDto(ReceitasDto receitas){
+        Receitas receita = ReceitasMapper.toEntity(receitas);
+        //Relacionar cada igrediente a receita
+        if (receitas.getIngredientes() != null){
+            receita.getIngredientes().forEach(i -> i.setReceita(receita));
+        }
+        Receitas salvar = receitasRepo.save(receita);
+        return ReceitasMapper.toDto(salvar);
+    }
 
     //listar
-    public List<Receitas> listarTodas(){
-      return receitasRepo.findAll();
+    public List<ReceitasDto> listarReceitasDto(){
+      List<Receitas> receitas = receitasRepo.findAll();
+      return receitas.stream().map(ReceitasMapper::toDto).collect(Collectors.toList());
     }
 
     //Busca por id
-    public Optional<Receitas> buscarId(int idReceita){
-        return receitasRepo.findById(idReceita);
+    public ReceitasDto buscarId(int idReceita){
+        Receitas receita = receitasRepo.findById(idReceita).orElseThrow(() -> new RuntimeException("Receita " +
+                "não encontrada com o id: " + idReceita));
+        return ReceitasMapper.toDto(receita);
     }
-
-    //Inserindo
-    public Receitas inserir(Receitas receitas){
-        return receitasRepo.save(receitas);
+    //atualizar
+    public ReceitasDto atualizar(int idReceita, ReceitasDto receitasDto) {
+        Receitas receita = receitasRepo.findById(idReceita).orElseThrow(() -> new RuntimeException("Receita " +
+                "não encontrada com o id: " + idReceita));
+        receita.setNome(receitasDto.getNome());
+        receita.setDescricao(receitasDto.getDescricao());
+        receita.setTempoDePreparo(receitasDto.getTempoDePreparo());
+        receita.getIngredientes().clear();
+        if (receitasDto.getIngredientes() != null) {
+            List<Ingrediente> novosIngredientes =
+                    receitasDto.getIngredientes().stream().map(IngredienteMapper::toEntity).peek(i -> i.setReceita(receita)).collect(Collectors.toList());
+            receita.getIngredientes().addAll(novosIngredientes);
+        }
+        Receitas atualizada = receitasRepo.save(receita);
+        return ReceitasMapper.toDto(atualizada);
     }
 
     //Deletando
     public void deletar(int idReceita){
+        if(!receitasRepo.existsById(idReceita)){
+            throw new RuntimeException("Receita não encontrada com o id: " + idReceita);
+        }
         receitasRepo.deleteById(idReceita);
+
     }
 }
